Combat/quadtree/other minor things
Entity- I worked on the entity class a lot more making subclasses like marine, maruader, worker. I also worked on the combat system and it works for basic combat.

QuadTree- worked on a quad tree to try and make enemy searching more efficient. Created the data structure and the units and enemies are both stored in a python list and a quad tree. Both are maintained and store the same data references. Quad tree is used when units need to find enemies, and units need to find other allies. For plain iterating use list its quicker.

Made minor changes to mathfuncs, input, and map classes.

*I think I kept all the Overlay changes from the last merge im going to check a little more after I merge this.