worked on shortest path movements
I worked on the shortest path algorithm. I don't know if it will be quick enough but it works. Still needs more work but the majority of the difficulty is done. Basically there's a graph of vertices and edges and when you select a unit and click it adds the unit and clicked position to the graph, runs the shortest path algorithm, and gives the unit a list of vertices to go to, and then removes them from the graph.

Adjacency List - added this data structure to hold the graphs
PriorityQueue - used this in the shortest path algorithm
BSP - used this to help see what vertices are visible on the graph given another vertex
MapGraph - holds the world graphs and the shortest path algorithm